refactor: enum들을 domain으로 옮기기
