+++ src/test/java/chess/domain/StartDecisionTest.java
@@ -1,6 +1,5 @@
 package chess.domain;
 
-import chess.controller.StartDecision;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
